neon/requirements.txt
#Use this command -> 'pip3 install -r requirements.txt --user' to install the dependencies automatically
getgauge
protobuf==3.20.0
selenium
websocket-client
requests
webdriver-manager
xmltodict
pyodbc
playwright
pylint
aenum
pythonnet >= 3.0.1
jsonschema
locust
locust-plugins
keyboard
flask
screeninfo
multipledispatch
setuptools
wheel
twine
configparser
pdoc
packaging
deepdiff
yapf
kombu
influxdb_client
wmi
retry-helper
websockets
flasgger